Add gp_backend_info() for runtime introspection/debugging
To debug into the master backend for a given Postgres session, you can
SELECT pg_backend_pid() and attach a debugger to the resulting process
ID. We currently have no corresponding function for the segment
backends, however -- developers have to read the output of `ps` and try
to correlate their connected session to the correct backends. This is
error-prone, especially if there are many sessions in flight.

gp_backend_info() is an attempt to fill this gap. Running

    SELECT * FROM gp_backend_info();

will return a table of the following format:

   id | type | content |   host    | port  |  pid
  ----+------+---------+-----------+-------+-------
   -1 | Q    |      -1 | pchampion | 25431 | 50430
    0 | w    |       0 | pchampion | 25432 | 50431
    1 | w    |       1 | pchampion | 25433 | 50432
    2 | w    |       2 | pchampion | 25434 | 50433

This allows developers to jump directly to the correct host and PID for
a given backend. This patch supports backends for writer gangs (type 'w'
in the table), reader gangs ('r'), master QD backend ('Q') and
master singleton readers ('R').

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>
Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>