Fix non-default collation fallbacks
For type NAMEOID, the default collation is C. Make sure various nodetypes are
aware of it.  Earlier only VAR node type took care of it. Extend it CONST and
OPEXPR.

A note about test case change in `select` ICG test. The test cases are supposed
to test for index scan on partial indexes. Earlier ORCA would fall back to
planner. After this fix, ORCA started to give sequential scan which was flaky
with explain analyze. So just like planner, we have turned off tablescan for
these tests. They are falling back to Planner for now.