add postmaster and all auxiliary processes to a same cgroup (#14256)
Currently, the postmaster process will be added to the parent cgroup, and all the
auxiliary processes, such as BgWriter, SysLogger, will be added to the cgroup of
user.slice, if we enable the resource group. We can not control the resource
usage of the cgroup of user.slice, and it's difficult to calculate the proportion
between the resource usage of the parent group and child group, the Linux Cgroup
document doesn't explain it either.

So this PR created a new control group, named "system_group", to control the
resource usage of the postmaster process and all other auxiliary processes.

And this PR uses the below principle:

When a process forks a child process, the new process is born into the
cgroup that the forking process belongs to at the time of the operation.
After exit, a process stays associated with the cgroup that it belonged to
at the time of exit until it's reaped;