refactor the original cgroup code and abstract the corresponding interface (#14343)
This PR is the second step in refactoring the resource group. The first one is #14256.

In this PR, we do not change any behavior of the resource group, we do not change the
interface exposed to the resource manager, it just abstracts all the fundamental functions
to the struct CGroupOpsRoutine, and use this abstract handler to manipulate the
underlying Linux Cgroup files, there are two purposes for this:

1. make the code more readable.
2. provides the base interface for Linux Cgroup v2.

The second one is our main motivation for doing this.

Of course, this is a relatively large change, so it's not all done, and more details need to be
fixed.