force the auto vacuum worker controlled by the system_group (#15573)
For the previous investigation, I neglected the fact that the AutoVacuum worker will be
executed by PostermasterMain, it will invoke InitResmanager to init the resource group if
it is enabled. And it'll use StartTransactionCommand to create a transaction to finish its
job, which could put the AutoVacuum worker process to the admin_group, not system_group,
so this PR is trying to fix it.